add correct pay tile for supporter
